+++ src/Commands/Base/ConnectOnline.cs
@@ -748,7 +748,7 @@ namespace PnP.PowerShell.Commands.Base
                 }
                 return false;
             }
-            catch(Exception ex)
+            catch
             {
                 return false;
             }